Abstract operation handlers for the reflection-based execution handler
