--- src/Microsoft.EntityFrameworkCore.Design/Design/Internal/DesignMarshalByRefObject.cs
@@ -1,14 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-
-namespace Microsoft.EntityFrameworkCore.Design.Internal
-{
-    public class DesignMarshalByRefObject
-#if NET451
-        : MarshalByRefObject
-#endif
-    {
-    }
-}
